# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/kitchens/{kitchenId}": {
      "get": {
        "operationId": "KitchensController_getOne",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KitchenDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Szczegóły kuchni i członkowie",
        "tags": [
          "kitchens"
        ]
      },
      "delete": {
        "operationId": "KitchensController_remove",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Usunięcie kuchni przez właściciela (kaskadowo członkowie, zaproszenia, produkty i partie)",
        "tags": [
          "kitchens"
        ]
      }
    },
...
    "/api/kitchens/{kitchenId}/products": {
      "get": {
        "operationId": "StockController_listProducts",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Katalog produktów kuchni",
        "tags": [
          "stock"
        ]
      },
      "post": {
        "operationId": "StockController_createProduct",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        },
        "summary": "Dodanie produktu do katalogu",
        "tags": [
          "stock"
        ]
      }
    },
    "/api/kitchens/{kitchenId}/products/{productId}": {
      "patch": {
        "operationId": "StockController_updateProduct",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        },
        "summary": "Aktualizacja produktu (m.in. purchaseMode)",
        "tags": [
          "stock"
        ]
      },
      "delete": {
        "operationId": "StockController_deleteProduct",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "confirmCascade",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Usunięcie produktu (kaskada partii po potwierdzeniu)",
        "tags": [
          "stock"
        ]
      }
    },
    "/api/kitchens/{kitchenId}/products/{productId}/configure-purchase": {
      "post": {
        "operationId": "StockController_configureProductPurchase",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigureProductPurchaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        },
        "summary": "Konfiguracja sposobu zakupu produktu (opakowania / dokładna ilość)",
        "tags": [
          "stock"
        ]
      }
    },
    "/api/kitchens/{kitchenId}/products/{productId}/nutrition": {
      "get": {
        "operationId": "StockController_getProductNutrition",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductNutritionDto"
                }
              }
            }
          }
        },
        "summary": "Wartości odżywcze produktu",
        "tags": [
          "stock"
        ]
      },
      "put": {
        "operationId": "StockController_upsertProductNutrition",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProductNutritionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductNutritionDto"
                }
              }
            }
          }
        },
        "summary": "Zapis wartości odżywczych produktu",
        "tags": [
          "stock"
        ]
      }
    },
    "/api/kitchens/{kitchenId}/products/{productId}/purchase-options": {
      "get": {
        "operationId": "StockController_listPurchaseOptions",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PurchaseOptionDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Opcje zakupu produktu",
        "tags": [
          "stock"
        ]
      },
      "post": {
        "operationId": "StockController_createPurchaseOption",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseOptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOptionDto"
                }
              }
            }
          }
        },
        "summary": "Dodanie opcji zakupu",
        "tags": [
          "stock"
        ]
      }
    },
    "/api/kitchens/{kitchenId}/products/{productId}/purchase-options/{optionId}": {
      "patch": {
        "operationId": "StockController_updatePurchaseOption",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "optionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePurchaseOptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOptionDto"
                }
              }
            }
          }
        },
        "summary": "Aktualizacja opcji zakupu",
        "tags": [
          "stock"
        ]
      },
      "delete": {
        "operationId": "StockController_deletePurchaseOption",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "optionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Usunięcie lub dezaktywacja opcji zakupu",
        "tags": [
          "stock"
        ]
      }
    },
    "/api/kitchens/{kitchenId}/stock-items": {
      "get": {
        "operationId": "StockController_listStock",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pantry",
                "fridge",
                "freezer",
                "other"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockItemDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Partie zapasów",
        "tags": [
          "stock"
        ]
      },
      "post": {
        "operationId": "StockController_createStock",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStockItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockItemDto"
                }
              }
            }
          }
        },
        "summary": "Dodanie partii zapasu",
        "tags": [
          "stock"
        ]
      }
    },
    "/api/kitchens/{kitchenId}/stock-items/{stockItemId}": {
      "patch": {
        "operationId": "StockController_updateStock",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stockItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStockItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockItemDto"
                }
              }
            }
          }
        },
        "summary": "Aktualizacja partii zapasu",
        "tags": [
          "stock"
        ]
      },
      "delete": {
        "operationId": "StockController_deleteStock",
        "parameters": [
          {
            "name": "kitchenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stockItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Usunięcie partii zapasu",
        "tags": [
          "stock"
        ]
      }
    },
...
  },
  "info": {
    "title": "Moja Kuchnia API",
    "description": "REST API aplikacji Moja Kuchnia",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
...
      "KitchenDetailsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KitchenMemberDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "members"
        ]
      },
...
      "ProductNutritionDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "baseQuantity": {
            "type": "string",
            "example": "100.000"
          },
          "baseUnit": {
            "type": "string",
            "enum": [
              "piece",
              "gram",
              "milliliter"
            ]
          },
          "kcal": {
            "type": "string",
            "example": "64.000"
          },
          "proteinGrams": {
            "type": "string"
          },
          "carbsGrams": {
            "type": "string"
          },
          "fatGrams": {
            "type": "string"
          },
          "fiberGrams": {
            "type": "string",
            "nullable": true
          },
          "saltGrams": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "productId",
          "baseQuantity",
          "baseUnit",
          "kcal",
          "proteinGrams",
          "carbsGrams",
          "fatGrams",
          "fiberGrams",
          "saltGrams",
          "updatedAt"
        ]
      },
      "PurchaseOptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "contentQuantity": {
            "type": "string",
            "example": "1000.000"
          },
          "contentUnit": {
            "type": "string",
            "enum": [
              "piece",
              "gram",
              "milliliter"
            ]
          },
          "isDefault": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "productId",
          "name",
          "contentQuantity",
          "contentUnit",
          "isDefault",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kitchenId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "normalizedName": {
            "type": "string"
          },
          "defaultUnit": {
            "type": "string",
            "enum": [
              "piece",
              "gram",
              "milliliter"
            ]
          },
          "purchaseMode": {
            "type": "string",
            "enum": [
              "unconfigured",
              "packaged",
              "exact"
            ]
          },
          "ean": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "description": "Starsze źródło zdjęcia (http lub data URL)."
          },
          "image": {
            "nullable": true,
            "description": "Zdjęcie z magazynu mediów; podpisane URL-e wygasają.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/MediaImageDto"
              }
            ]
          },
          "nutrition": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductNutritionDto"
              }
            ]
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "purchaseOptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseOptionDto"
            }
          }
        },
        "required": [
          "id",
          "kitchenId",
          "name",
          "normalizedName",
          "defaultUnit",
          "purchaseMode",
          "ean",
          "imageUrl",
          "image",
          "nutrition",
          "category",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Mleko"
          },
          "defaultUnit": {
            "type": "string",
            "enum": [
              "piece",
              "gram",
              "milliliter"
            ],
            "example": "milliliter"
          },
          "ean": {
            "type": "string",
            "nullable": true,
            "example": "5901234123457",
            "description": "Kod EAN/GTIN (8, 12, 13 lub 14 cyfr)."
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "description": "URL http(s) albo data URL obrazu (jpeg/png/webp/gif)."
          },
          "category": {
            "type": "string",
            "nullable": true,
            "example": "Nabiał"
          }
        },
        "required": [
          "name",
          "defaultUnit"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "purchaseMode": {
            "type": "string",
            "enum": [
              "unconfigured",
              "packaged",
              "exact"
            ]
          }
        }
      },
...
      "ConfigureProductPurchaseDto": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "unconfigured",
              "packaged",
              "exact"
            ],
            "description": "packaged wymaga pierwszej opcji; exact nie wymaga opcji; unconfigured czyści tryb bez usuwania opcji."
          },
          "option": {
            "$ref": "#/components/schemas/ConfigureProductPurchaseOptionDto"
          }
        },
        "required": [
          "mode"
        ]
      },
      "UpsertProductNutritionDto": {
        "type": "object",
        "properties": {
          "baseQuantity": {
            "type": "string",
            "example": "100.000",
            "description": "Ilość odniesienia w jednostce bazowej produktu."
          },
          "baseUnit": {
            "type": "string",
            "enum": [
              "piece",
              "gram",
              "milliliter"
            ],
            "example": "milliliter"
          },
          "kcal": {
            "type": "string",
            "example": "64.000"
          },
          "proteinGrams": {
            "type": "string",
            "example": "3.200"
          },
          "carbsGrams": {
            "type": "string",
            "example": "4.700"
          },
          "fatGrams": {
            "type": "string",
            "example": "3.600"
          },
          "fiberGrams": {
            "type": "string",
            "nullable": true,
            "example": "0.000"
          },
          "saltGrams": {
            "type": "string",
            "nullable": true,
            "example": "0.100"
          }
        },
        "required": [
          "baseQuantity",
          "baseUnit",
          "kcal",
          "proteinGrams",
          "carbsGrams",
          "fatGrams"
        ]
      },
      "CreatePurchaseOptionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Karton 1 l"
          },
          "contentQuantity": {
            "type": "string",
            "example": "1000.000"
          },
          "contentUnit": {
            "type": "string",
            "enum": [
              "piece",
              "gram",
              "milliliter"
            ]
          },
          "isDefault": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "name",
          "contentQuantity",
          "contentUnit"
        ]
      },
      "UpdatePurchaseOptionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Karton 1 l"
          },
          "contentQuantity": {
            "type": "string",
            "example": "1000.000"
          },
          "contentUnit": {
            "type": "string",
            "enum": [
              "piece",
              "gram",
              "milliliter"
            ]
          },
          "isDefault": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}